# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/duties": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a list of all duties",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "duties"
                ],
                "summary": "List all duties",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Duty"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new duty in the system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "duties"
                ],
                "summary": "Create a new duty",
                "parameters": [
                    {
                        "description": "Duty",
                        "name": "duty",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Duty"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Duty"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/duties/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a duty by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "duties"
                ],
                "summary": "Get a duty by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Duty ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Duty"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update a duty by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "duties"
                ],
                "summary": "Update a duty",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Duty ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Duty",
                        "name": "duty",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Duty"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Duty"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a duty by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "duties"
                ],
                "summary": "Delete a duty",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Duty ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Conscript": {
            "description": "Conscript is a user entity used for authentication and as a foreign key in other models. It includes unique registry and username fields, a password (should be hashed in production), and belongs to a department. Timestamps are managed by Gorm.",
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "department": {
                    "$ref": "#/definitions/models.Department"
                },
                "departmentID": {
                    "type": "integer"
                },
                "firstName": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "lastName": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "registryNumber": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "models.ConscriptDuty": {
            "description": "ConscriptDuty is the join table for conscripts and duties, with assignment period and timestamps. Composite primary key: conscript_id, duty_id.",
            "type": "object",
            "properties": {
                "conscriptID": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "dutyID": {
                    "type": "integer"
                },
                "endTime": {
                    "type": "string"
                },
                "startTime": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "models.Department": {
            "description": "Department is a unique grouping for conscripts and services. It is referenced by conscripts and services, and includes a unique label. Timestamps are managed by Gorm.",
            "type": "object",
            "properties": {
                "conscripts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Conscript"
                    }
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "label": {
                    "type": "string"
                },
                "services": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Service"
                    }
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "models.Duty": {
            "description": "Duty is a task or responsibility assigned to conscripts, linked to a service, and can be assigned to many conscripts. Timestamps are managed by Gorm.",
            "type": "object",
            "properties": {
                "conscriptDuties": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ConscriptDuty"
                    }
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "label": {
                    "type": "string"
                },
                "service": {
                    "$ref": "#/definitions/models.Service"
                },
                "serviceID": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "models.Service": {
            "description": "Service is a grouping of duties within a department. Label is unique. Timestamps are managed by Gorm.",
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "department": {
                    "$ref": "#/definitions/models.Department"
                },
                "departmentID": {
                    "type": "integer"
                },
                "duties": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Duty"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "label": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        }
    }
}